Added description of mp3 files with the content of unsychronised lyrics and comment tags.
Up until now the "desc" attribute was only added for podcast items and
filled with a description of that item. Now the "desc" attribute is filled
for all mp3 files with the merged content of ULT/USLT tags and the COM/COMM tags.
For podcasts the item description is merged into the "desc" attribute as well.